+++ ui/src/components/flows/FlowTriggers.vue
@@ -237,7 +237,7 @@
     import TriggerAvatar from "./TriggerAvatar.vue";
 
     import KestraFilter from "../filter/KestraFilter.vue";
-    import Empty from "../layout/Empty.vue";
+    import Empty from "../layout/empty/Empty.vue";
 </script>
 
 <script>
