+++ ui/src/components/layout/empty/assets/visuals/announcements.png
