+++ ui/src/components/layout/empty/assets/visuals/apps.png
