+++ ui/src/components/layout/empty/assets/visuals/concurrency_executions.png
