+++ ui/src/components/layout/empty/assets/visuals/concurrency_limit.png
