+++ ui/src/components/layout/empty/assets/visuals/dependencies.png
