+++ ui/src/components/layout/empty/assets/visuals/plugins.png
