+++ ui/src/components/layout/empty/assets/visuals/triggers.png
