+++ ui/src/components/layout/empty/images.ts
@@ -1,0 +1,17 @@
+import announcements from "./assets/visuals/announcements.png";
+import apps from "./assets/visuals/apps.png";
+import concurrency_executions from "./assets/visuals/concurrency_executions.png";
+import concurrency_limit from "./assets/visuals/concurrency_limit.png";
+import dependencies from "./assets/visuals/dependencies.png";
+import plugins from "./assets/visuals/plugins.png";
+import triggers from "./assets/visuals/triggers.png";
+
+export const images: Record<string, string> = {
+    announcements,
+    apps,
+    concurrency_executions,
+    concurrency_limit,
+    dependencies,
+    plugins,
+    triggers,
+};
